requirements.txt
asgiref==3.7.2
boto3==1.28.71
botocore==1.31.71
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jmespath==1.0.1
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
